+++ src/System.Private.CoreLib/shared/System/Runtime/Intrinsics/Vector256.cs
@@ -1770,7 +1770,7 @@ namespace System.Runtime.Intrinsics
             }
 
             ref T e0 = ref Unsafe.As<Vector256<T>, T>(ref vector);
-            return Unsafe.Add(ref e0, index);
+            return Unsafe.Add(ref e0, (uint)index);
         }
 
         /// <summary>Creates a new <see cref="Vector256{T}" /> with the element at the specified index set to the specified value and the remaining elements set to the same value as that in the given vector.</summary>
@@ -1794,7 +1794,7 @@ namespace System.Runtime.Intrinsics
 
             Vector256<T> result = vector;
             ref T e0 = ref Unsafe.As<Vector256<T>, T>(ref result);
-            Unsafe.Add(ref e0, index) = value;
+            Unsafe.Add(ref e0, (uint)index) = value;
             return result;
         }
