+++ src/System.Private.CoreLib/shared/System/Runtime/Intrinsics/Vector64.cs
@@ -619,7 +619,7 @@ namespace System.Runtime.Intrinsics
             }
 
             ref T e0 = ref Unsafe.As<Vector64<T>, T>(ref vector);
-            return Unsafe.Add(ref e0, index);
+            return Unsafe.Add(ref e0, (uint)index);
         }
 
         /// <summary>Creates a new <see cref="Vector64{T}" /> with the element at the specified index set to the specified value and the remaining elements set to the same value as that in the given vector.</summary>
@@ -642,7 +642,7 @@ namespace System.Runtime.Intrinsics
 
             Vector64<T> result = vector;
             ref T e0 = ref Unsafe.As<Vector64<T>, T>(ref result);
-            Unsafe.Add(ref e0, index) = value;
+            Unsafe.Add(ref e0, (uint)index) = value;
             return result;
         }
